update and reuse preprocess_inputs
