+++ optimum/intel/openvino/modeling_base_seq2seq.py
@@ -87,7 +87,7 @@ class OVBaseModelForSeq2SeqLM(OVBaseModel):
             # some model configs may have issues with loading without parameters initialization
             try:
                 misplaced_generation_parameters = self.config._get_non_default_generation_parameters()
-            except Exception:
+            except KeyError:
                 misplaced_generation_parameters = {}
             if len(misplaced_generation_parameters) > 0:
                 logger.warning(
